+++ index-dev.js
@@ -5,5 +5,11 @@ module.exports.Responsive = require("./lib/ResponsiveReactGridLayout").default;
 module.exports.Responsive.utils = require("./lib/responsiveUtils");
 module.exports.WidthProvider =
   require("./lib/components/WidthProvider").default;
-module.exports.CrossGridProvider =
-  require("./lib/CrossGridProvider").default;
+
+// Drag and drop exports
+module.exports.DragDropProvider =
+  require("./lib/DragDropProvider").default;
+module.exports.DragDropContext =
+  require("./lib/DragDropContext").default;
+module.exports.Droppable =
+  require("./lib/components/Droppable").default;